+++ src/components/__tests__/MultiNodeBucket.test.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import { UnconnectedOrdinalBinBucket as OrdinalBinBucket } from '../OrdinalBinBucket';
+import { UnconnectedMultiNodeBucket as MultiNodeBucket } from '../MultiNodeBucket';
 import { NO_SCROLL } from '../../behaviours/DragAndDrop/DragManager';
 
 jest.mock('../../containers/Node');
@@ -11,7 +11,7 @@ describe('OrdinalBinBucket', () => {
   let bucket;
 
   beforeEach(() => {
-    bucket = mount(<OrdinalBinBucket nodes={[{}]} />);
+    bucket = mount(<MultiNodeBucket nodes={[{}]} />);
   });
 
   it('renders connected node items', () => {
