+++ src/components/index.js
@@ -8,7 +8,7 @@ export { default as Dialog } from './Dialog';
 export { default as Modal } from './Modal';
 export { default as Edge } from './Edge';
 export { default as NodeList } from './NodeList';
-export { default as OrdinalBinBucket } from './OrdinalBinBucket';
+export { default as MultiNodeBucket } from './MultiNodeBucket';
 export { default as Panels } from './Panels';
 export { default as Panel } from './Panel';
 export { default as DropZone } from './DropZone';
